requirements.txt
Click==7.0
dlib==19.17.0
face-recognition==1.2.3
face-recognition-models==0.3.0
numpy==1.16.4
Pillow==6.0.0
